backend/pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roboyard-control-backend"
version = "0.1.0"
description = "RoboYard Control FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.21.0",
  "asyncpg>=0.30.0",
  "celery[redis]>=5.5.3",
  "email-validator>=2.3.0",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "passlib[bcrypt]>=1.7.4",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "pydantic-settings>=2.11.0",
  "PyJWT>=2.10.1",
  "python-json-logger>=3.3.0",
  "python-multipart>=0.0.20",
  "redis>=6.4.0",
  "sqlalchemy[asyncio]>=2.0.43",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "asgi-lifespan>=2.1.0",
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
